fix(web): scope the canvas resize observer to its own editor
The observer found its container with
`document.querySelector(".excalidraw")`.
That selector is not unique even with a single editor mounted:
Excalidraw's
portal helper does `classList.add("excalidraw", ...)` on the divs it
appends to
`document.body` for modals and tooltips, so the class matches those too.
The
lookup worked only because the editor happens to precede them in
document
order. Had a portal won, `container.querySelector("canvas")` would
return null,
the width comparison would return early every time, and the fix would be
dead
with nothing to show for it.

`ExcalidrawImperativeAPI` exposes no container accessor to scope the
query with
-- `dist/types/excalidraw/types.d.ts` gives `id`, `refresh` and the
scene
methods, nothing pointing at the DOM. So the effect moved into
`Whiteboard.tsx`
instead, which already renders the wrapper div and can hold a ref to it.
That
wrapper is the element a pane change actually resizes, so this is both
scoped
to the right instance and one lookup shorter. It stays gated on the API
callback, now via local state, because the canvases exist only once the
dynamically imported editor has mounted.

The callback also re-observes the canvases on each run, which picks up
the
new-element canvas Excalidraw mounts mid-stroke. A repeat `observe()` on
an
already-observed target is a no-op, so this costs nothing: measured two
callbacks per toggle, one to detect the disagreement and one to confirm
the
canvas caught up, with no further delivery.
