+++ apps/web/src/components/whiteboard/workspace-layout/useExcalidrawScene.ts
@@ -9,37 +9,6 @@ export function useExcalidrawScene(initialScene: unknown) {
     setApi((currentAPI) => (currentAPI === nextAPI ? currentAPI : nextAPI));
   }, []);
 
-  // Excalidraw re-measures its canvas only on window resize, not container
-  // resize. Pane changes (sidebar open/close, divider drag) move the container
-  // but don't fire a resize, leaving blank space. Dispatching a synthetic
-  // resize event forces Excalidraw to re-measure. Chose this over
-  // `api.refresh()` because refresh() recomputes scroll offsets but leaves
-  // canvas dimensions untouched — measured both: refresh() left a 1524px
-  // canvas stuck in a 1908px container, resize event fixed it.
-  useEffect(() => {
-    if (!api) return;
-    const container = document.querySelector(".excalidraw");
-    if (!(container instanceof HTMLElement)) return;
-
-    // Compare container width vs canvas width. ResizeObserver alone fires
-    // before Excalidraw finishes its initial sizing, so the correction gets
-    // overwritten. Checking both catches whichever side moves last.
-    const observer = new ResizeObserver(() => {
-      const width = container.getBoundingClientRect().width;
-      const canvas = container.querySelector("canvas");
-      if (!canvas || Math.abs(canvas.getBoundingClientRect().width - width) < 1) return;
-      // Terminates: the event makes Excalidraw re-measure to `width`, after
-      // which the two agree and this returns above.
-      window.dispatchEvent(new Event("resize"));
-    });
-
-    observer.observe(container);
-    // The canvases too, so Excalidraw sizing itself late re-runs the check.
-    for (const canvas of container.querySelectorAll("canvas")) observer.observe(canvas);
-
-    return () => observer.disconnect();
-  }, [api]);
-
   useEffect(() => {
     if (!api) return;
     const scene = resolveDiagramScene(initialScene);
